narrative_engine/memory/requirements-vector.txt
faiss-cpu
numpy
# openai  # Removed - using Ollama instead 